+++ adminUsers.py
@@ -3,7 +3,8 @@ import shelve
 from flask import flash, Blueprint, render_template, request, session, redirect, url_for
 from functions import flashFormErrors, goBack, adminAccess
 from classes.User import User
-from forms import editUserForm, searchUsersForm, changeUserPasswordForm, addUserForm
+from classes.Address import Address
+from forms import editUserForm, searchUsersForm, changeUserPasswordForm, addUserForm, addressForm
 
 adminUsers = Blueprint("adminUsers", __name__)
 
@@ -47,6 +48,43 @@ def editPassword(email):
 
     with shelve.open("users") as users:
         return render_template("admin/editPassword.html", user=users[email], form=form)
+
+@adminUsers.route("/admin/users/edit/address/<email>", methods=['GET', 'POST'])
+@adminAccess
+def viewAddresses(email):
+
+    with shelve.open("users") as users:
+        try:
+            with shelve.open("addresses") as allAddresses:
+                addresses = allAddresses[email]
+        except KeyError:
+            addresses = []
+            flash("No addresses added yet for this account.", category="warning")
+        finally:
+            return render_template("admin/editAddress.html", user=users[email], addresses=addresses)
+
+@adminUsers.route("/admin/users/add/address/<email>", methods=['GET', 'POST'])
+@adminAccess
+def addAddress(email):
+    form = addressForm(request.form)
+
+    if request.method == "POST" and form.validate():
+        print("Add address")
+        address = Address(form.name.data, form.location.data)
+
+        with shelve.open("addresses", writeback=True) as addresses:
+            if email not in addresses.keys():
+                addresses[email] = [address]
+            else:
+                addresses[email].append(address)
+
+            flash("Address has been successfully added")
+            return redirect(url_for("adminUsers.viewAddresses", email=email))
+    else:
+        flashFormErrors("Unable to add address", form.errors)
+
+    with shelve.open("users") as users:
+        return render_template("admin/addAddress.html", user=users[email], form=form)
 
 @adminUsers.route("/admin/users/add", methods=['GET', 'POST'])
 @adminAccess