+++ forms.py
@@ -129,3 +129,16 @@ class addUserForm(Form):
     def validate_birthdate_field(form, field):
         if form.birthday.data > datetime.date.today():
             raise ValidationError("Invalid birthday, date cannot be in the future")
+
+
+class addressForm(Form):
+    name = StringField("Address Name", [
+        validators.Length(1, 64, message="Address name must be between 1 to 64 characters"),
+        validators.DataRequired(message="Address name is required")
+    ])
+    location = StringField("Location", [
+        validators.Length(16, 256, message="Location must be between 16 to 256 characters"),
+        validators.DataRequired(message="Location is required")
+    ])
+
+    submit = SubmitField("Add Address")